SOLR-17518: Restore compilation backward compability of RequestWriter. (#3288)
